hashed vmid and processname
