+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/VmMetadataApiHandlerTest.cs
@@ -18,6 +18,7 @@ namespace Microsoft.Azure.Cosmos
     using Microsoft.VisualStudio.TestTools.UnitTesting;
     using Moq;
     using Newtonsoft.Json;
+    using Util;
 
     [TestClass]
     public class VmMetadataApiHandlerTest
@@ -56,7 +57,7 @@ namespace Microsoft.Azure.Cosmos
             VmMetadataApiHandler.TryInitialize(cosmoshttpClient);
 
             await Task.Delay(2000);
-            Assert.AreEqual("vmId:d0cb93eb-214b-4c2b-bd3d-cc93e90d9efd", VmMetadataApiHandler.GetMachineId());
+            Assert.AreEqual($"vmId:{CosmosUtils.ComputeHash("d0cb93eb-214b-4c2b-bd3d-cc93e90d9efd")}", VmMetadataApiHandler.GetMachineId());
             Assert.AreEqual(VmMetadataApiHandler.GetMachineRegion(), "eastus");
         }
 
@@ -88,7 +89,7 @@ namespace Microsoft.Azure.Cosmos
         [TestMethod]
         public async Task GetHashedMachineNameAsMachineIdTest()
         {
-            string expectedMachineId = "hashedMachineName:" + VmMetadataApiHandler.ComputeHash(Environment.MachineName);
+            string expectedMachineId = "hashedMachineName:" + CosmosUtils.ComputeHash(Environment.MachineName);
 
             static Task<HttpResponseMessage> sendFunc(HttpRequestMessage request, CancellationToken cancellationToken) { throw new Exception("error while making API call"); };
 
@@ -105,7 +106,7 @@ namespace Microsoft.Azure.Cosmos
         [TestMethod]
         public void ComputeHashTest()
         {
-            string hashedValue = VmMetadataApiHandler.ComputeHash("abc");
+            string hashedValue = CosmosUtils.ComputeHash("abc");
             Assert.AreEqual("bf1678ba-018f-eacf-4141-40de5dae2223", hashedValue);
         }
 
@@ -128,7 +129,7 @@ namespace Microsoft.Azure.Cosmos
             Assert.AreEqual("AzurePublicCloud", metadata.Compute.AzEnvironment);
             Assert.AreEqual("Linux", metadata.Compute.OSType);
             Assert.AreEqual("Standard_D2s_v3", metadata.Compute.VMSize);
-            Assert.AreEqual("vmId:d0cb93eb-214b-4c2b-bd3d-cc93e90d9efd", metadata.Compute.VMId);
+            Assert.AreEqual($"vmId:{CosmosUtils.ComputeHash("d0cb93eb-214b-4c2b-bd3d-cc93e90d9efd")}", metadata.Compute.VMId);
         }
 
         [TestMethod]
